chore(exception): change name and move package
