+++ src/main/java/lotto/exception/administrator/LottoNumbersMustBeSixException.java
@@ -1,0 +1,7 @@
+package lotto.exception.administrator;
+
+public class LottoNumbersMustBeSixException extends IllegalArgumentException {
+    public LottoNumbersMustBeSixException() {
+        super("[ERROR] 로또 번호는 6개여야 합니다.");
+    }
+}
