+++ src/main/java/lotto/exception/user/NotEnoughMoneyException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.exception.user;
 
 public class NotEnoughMoneyException extends IllegalArgumentException {
     public NotEnoughMoneyException() {
